+++ src/commands/inherit_class.js
@@ -29,7 +29,7 @@ async function inheritClass(uri) {
     }
 
     let interfaceName = textFile.substring(indexStart, indexEnd).replace(stringSearch, ' ').replace(' {', '');
-    interfaceName = interfaceName.trimEnd();
+    interfaceName = interfaceName.trimRight();
     const configType = vscode.workspace.getConfiguration("generate").get("template.type");
     let implementationName = interfaceName.trim();
     if (configType === 'I Prefix') {